refactor: product 패키지 분리
