+++ src/main/java/kitchenpos/domain/MenuProduct.java
@@ -8,6 +8,7 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import kitchenpos.domain.product.Product;
 import kitchenpos.support.domain.BaseEntity;
 import kitchenpos.support.money.Money;
