+++ src/main/java/kitchenpos/domain/product/ProductRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.dao;
+package kitchenpos.domain.product;
 
 import java.util.List;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.product.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {
