+++ src/main/java/kitchenpos/dto/MenuRequest.java
@@ -9,7 +9,7 @@ import java.util.List;
 import java.util.Map;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.Product;
+import kitchenpos.domain.product.Product;
 
 public class MenuRequest {
