+++ src/main/java/kitchenpos/dto/product/ProductRequest.java
@@ -1,22 +1,17 @@
-package kitchenpos.dto;
+package kitchenpos.dto.product;
 
 import java.math.BigDecimal;
-import kitchenpos.domain.Product;
 
-public class ProductResponse {
+public class ProductRequest {
 
     private Long id;
     private String name;
     private BigDecimal price;
 
-    public ProductResponse(Long id, String name, BigDecimal price) {
+    public ProductRequest(Long id, String name, BigDecimal price) {
         this.id = id;
         this.name = name;
         this.price = price;
-    }
-
-    public static ProductResponse from(Product product) {
-        return new ProductResponse(product.getId(), product.getName(), product.getPrice().getAmount());
     }
 
     public Long getId() {
