+++ src/test/java/kitchenpos/fixture/ProductFixture.java
@@ -1,6 +1,6 @@
 package kitchenpos.fixture;
 
-import kitchenpos.domain.Product;
+import kitchenpos.domain.product.Product;
 import kitchenpos.support.money.Money;
 
 public class ProductFixture {
